reset screen offset of top line if it exceeds the number of screens
Some operations such as changing the tab size may change the number of
screens needed to display a line and may therefore result in the current
screen offset to exceed this number of screens.
This could in turn result in the entire screen not being populated properly.

This patch detects and corrects this situation while populating the
screen after reformatting.

Problem reported and fixed by patrick keshishian <pkeshish@gmail.com>